move: move download_list_type.py into download_mode.py
